refactor: remove separate singleton-based theming package
