+++ src/modules/garden-theme/index.ts
@@ -1,3 +1,3 @@
-export { setupGardenTheme } from "./setupGardenTheme";
+export { createTheme } from "./createTheme";
 export { ThemeProviders } from "./ThemeProviders";
 export { useModalContainer } from "./modal-container/useModalContainer";
