--- src/modules/theming/styled.d.ts
@@ -1,6 +1,0 @@
-import type { IGardenTheme } from "@zendeskgarden/react-theming";
-import "styled-components";
-
-declare module "styled-components" {
-  export interface DefaultTheme extends IGardenTheme {}
-}
